Fix type safety issues flagged in code review
- main.dart: add .toInt()/.toDouble() on clamped values; compare fields
  before copyWith so setState is skipped when nothing changed
- CameraController.kt: convert Float greenEven/greenOdd to Double before
  averaging to avoid Float/Double type mismatch
- Messages.g.swift: keep PigeonError.details as Any? instead of force-casting
  to String? via nilOrValue (3 sites) to prevent runtime crashes

+++ packages/cambrian_camera/android/src/main/kotlin/com/cambrian/camera/CameraController.kt
@@ -1299,7 +1299,7 @@ class CameraController(
                         exposureTimeNs = lastKnownExposureTimeNs,
                         focusDistanceDiopters = focusDist?.toDouble(),
                         wbGainR = wbGains?.red?.toDouble(),
-                        wbGainG = wbGains?.let { (it.greenEven + it.greenOdd) / 2.0 },
+                        wbGainG = wbGains?.let { (it.greenEven.toDouble() + it.greenOdd.toDouble()) / 2.0 },
                         wbGainB = wbGains?.blue?.toDouble(),
                     )
                     mainHandler.post { flutterApi.onFrameResult(handle, frameResult) {} }
